+++ tol_lab_share/messages/consumed/traction_to_warehouse/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["TractionToWarehouseMessage"]
 
-from .root import Root as TractionToWarehouseMessage
+from .traction_to_mlwh_aliquot import Root as TractionToWarehouseMessage
